% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classification_dataset_names}
\alias{classification_dataset_names}
\title{Names of available classification datasets}
\format{
An object of class \code{character} of length 162.
}
\source{
\url{https://github.com/EpistasisLab/pmlb}
}
\usage{
classification_dataset_names
}
\description{
A list of the names of available classification datasets
}
\keyword{datasets}
