% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{classification_datasets}
\alias{classification_datasets}
\title{Classification datasets}
\usage{
classification_datasets()
}
\value{
A character vector of classification dataset names.
}
\description{
Classification datasets
}
\examples{
if (interactive()) {
  sample(classification_datasets(), 10)
}
}
