% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{compute_imbalance}
\alias{compute_imbalance}
\title{Computes imbalance value for a given dataset.}
\usage{
compute_imbalance(target_col)
}
\arguments{
\item{target_col}{Factor or character vector of target column.}
}
\value{
A value of imbalance metric,
where zero means that the dataset is perfectly balanced
and the higher the value, the more imbalanced the dataset.
}
\description{
Computes imbalance value for a given dataset.
}
