% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{dataset_names}
\alias{dataset_names}
\title{All available datasets}
\usage{
dataset_names()
}
\value{
A character vector of all dataset names.
}
\description{
All available datasets
}
\examples{
if (interactive()) {
  sample(dataset_names(), 10)
}
}
