% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest.R
\name{get_type}
\alias{get_type}
\title{Get type/class of given vector.}
\usage{
get_type(x, include_binary = FALSE)
}
\arguments{
\item{x}{Input vector.}

\item{include_binary}{Boolean.
Whether binary should be counted separately from categorical.}
}
\value{
Type/class of `x`.
}
\description{
Get type/class of given vector.
}
