% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{pmlb_metadata}
\alias{pmlb_metadata}
\title{Get Summary Statistics}
\usage{
pmlb_metadata()
}
\value{
A list containing summary_stats, dataset_names, classification_dataset_names, and regression_dataset_names
}
\description{
Downloads and processes the latest summary statistics from PMLB.
Use cache if available.
}
\examples{
if (interactive()) {
  sample(pmlb_metadata()$dataset_names, 10)
}
}
