% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmlbr-package.R
\docType{package}
\name{pmlbr-package}
\alias{pmlbr}
\alias{pmlbr-package}
\title{pmlbr: Interface to the Penn Machine Learning Benchmarks Data Repository}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Check available classification and regression data sets from the PMLB repository and download them. The PMLB repository (\url{https://github.com/EpistasisLab/pmlbr}) contains a curated collection of data sets for evaluating and comparing machine learning algorithms. These data sets cover a range of applications, and include binary/multi-class classification problems and regression problems, as well as combinations of categorical, ordinal, and continuous features. There are currently over 150 datasets included in the PMLB repository.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EpistasisLab/pmlbr}
  \item Report bugs at \url{https://github.com/EpistasisLab/pmlbr/issues}
}

}
\author{
\strong{Maintainer}: Trang Le \email{grixor@gmail.com} (https://trang.page/)

Authors:
\itemize{
  \item makeyourownmaker \email{makeyourownmaker@gmx.com} (https://github.com/makeyourownmaker)
  \item Jason Moore \email{jhmoore@upenn.edu} (http://www.epistasisblog.org/)
}

Other contributors:
\itemize{
  \item University of Pennsylvania [copyright holder]
}

}
\keyword{internal}
