% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regression_dataset_names}
\alias{regression_dataset_names}
\title{Names of available regression datasets}
\format{
An object of class \code{character} of length 122.
}
\source{
\url{https://github.com/EpistasisLab/pmlb}
}
\usage{
regression_dataset_names
}
\description{
A list of the names of available regression datasets
}
\keyword{datasets}
