% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{regression_datasets}
\alias{regression_datasets}
\title{Regression datasets}
\usage{
regression_datasets()
}
\value{
A character vector of regression dataset names.
}
\description{
Regression datasets
}
\examples{
if (interactive()) {
  sample(regression_datasets(), 10)
}
}
